fix(ci): MSVC C4365 signed/unsigned 불일치 수정 — 신규 cpp 파일 4건
xpe_calib_generate_gain.cpp:
- fit_polynomial_ls: int32_t 루프 변수 → size_t, aug[][] 인덱스 정합
- 역방향 루프 size_t 관용구 적용 (i = sM; i-- > 0)
- gain_maps/headers/coeff_array 벡터 생성자에 size_t 캐스트
- FUNC-027 픽셀 루프 내 int32_t 인덱스 → size_t 전환

xpe_verify_metrics.cpp:
- compute_flatness: hist 벡터 생성자 bins → static_cast<size_t>(bins)

+++ modules/preprocess/src/xpe_verify_metrics.cpp
@@ -78,7 +78,7 @@ namespace {
         if (max_val <= min_val) return 1.0; // All values identical → perfectly flat
 
         // Build histogram
-        std::vector<int> hist(bins, 0);
+        std::vector<int> hist(static_cast<size_t>(bins), 0);
         double bin_width = (max_val - min_val) / bins;
 
         for (double v : values) {
